引入agql查询Steam服务器信息

+++ src/main/java/com/jinlink/modules/game/entity/GameOnlineStatistics.java
@@ -36,5 +36,5 @@ public class GameOnlineStatistics extends BaseEntity {
     /**
      * 社区在线人数
      */
-    private Integer communityPlay;
+    private Long communityPlay;
 }
